requirements.txt
asgiref==3.2.7
Django==3.0.6
django-ckeditor==5.9.0
django-cleanup==4.0.0
django-js-asset==1.2.2
django-taggit==1.2.0
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
